+++ Changelog.md
@@ -5,14 +5,15 @@
 - mkvtoolnix 48
 - ec3 file extension support for eac3 files.
 - R210/V210 video output using ffmpeg.
-- Include beta versions for update checker and show changelog (Dendraspis)
+- The final assistant tip supports SHIFT key to add the job at the top
+  of the job list and CTRL to prevent showing the jobs dialog.
+  It's documented via tooltip, right-click on Next button. (Dendraspis)
 
 
 2.1.3.7 Beta
 ============
 
 - AviSynth soft link fix.
-- DGIndex re-enabled by default for VOB/MPG.
 
 
 2.1.3.6 Beta